Change AUTO mode to HEAT mode
HomeAssistant 2025.10.0 doesn't display a temperature setting option if
the device is in auto mode, which makes sense, since the mode is meant
to describe the device dynamically changing the temperature. Water
heaters can only heat up to a certain temperature, so to reflect that,
change the mode to heat mode.

Signed-off-by: Saikrishna Arcot <saiarcot895@gmail.com>